backend/requirements.txt
flask
flask-cors
pandas
prophet==1.1.5
gunicorn
flask-sqlalchemy
flask-jwt-extended
psycopg2-binary
werkzeug
flask-migrate
numpy==1.24.3
scikit-learn
flask-mail
python-dotenv
apscheduler
requests
textblob
flask-socketio